Clean the code and remove some useless methods
