+++ runbot_github/models/github.py
@@ -1,0 +1,24 @@
+from openerp.addons.runbot_multiple_hosting import hosting
+
+class GithubHosting(hosting.Hosting):
+    API_URL = 'https://api.github.com'
+    URL = 'https://github.com'
+
+    def __init__(self, credentials):
+        token = (credentials, 'x-oauth-basic')
+        super(GithubHosting, self).__init__(token)
+
+        self.session.headers.update({'Accept': 'application/vnd.github.she-hulk-preview+json'})
+
+    @classmethod
+    def get_branch_url(cls, owner, repository, branch):
+        return cls.get_url('/%s/%s/tree/%s', owner, repository, branch)
+
+    @classmethod
+    def get_pull_request_url(cls, owner, repository, pull_number):
+        return cls.get_url('/%s/%s/pull/%s', owner, repository, pull_number)
+
+    def get_pull_request(self, owner, repository, pull_number):
+        url = self.get_api_url('/repos/%s/%s/pulls/%s' % (owner, repository, pull_number))
+        response = self.session.get(url)
+        return response.json()
